diyabettakip/requirements.txt
PyQt5>=5.15.0
mysql-connector-python>=8.0.0
matplotlib>=3.5.0
numpy>=1.21.0
